SyncContext: use AuthProvider
When running a local sync, the syncURL/username/password are not meant
for the sync and cannot be used if they refer to an AuthProvider which
cannot return plain username/password.

In all other cases, this may or may not work, so at least try it instead
of hard-coding the IdentityProviderCredentials.