GNOME: work around GNOME keyring communication problem
It seems that sometimes setting up a session with GNOME keyring fails such
that all further communication leads to decoding problem.

There is an internal method to reset the session, but it cannot be called
directly. As a workaround, fake the death of the GNOME keyring daemon
and thus trigger a reconnect when retrying the GNOME keyring access.