compatibility hack: support libical.so.1
libical 1.0 changed the library name to libical.so.1, causing dependency
problems becaus syncevolution.org binaries required libical.so.0.

It turned out that the relevant parts of the API and ABI had not changed,
so simply allowing both versions of the library is sufficient.

In addition, a detection of libical.so.1 gets added, just in case.