signon: fix compilation with nothing enabled
When neither gSSO nor UOA were enabled, the provideruoa.so was
still enabled and failed to link because its LDFLAGS were not
set. It shouldn't have been enabled at all.

While we are at it, allow both to be enabled in the build rules,
even though the code doesn't support it (symbol clashes).