local sync: avoid redundant ConfigNode instances
When configuring a local sync on the target side, the code in
SyncConfig::getSyncSourceNodes() created redundant ConfigNode instances.  This
was pointed out by cppcheck as redundant variable assignments.  The code which
created these additional instances may have had side-effects, but it doesn't
look like these were desired and/or relevant.