GNOME Bluetooth: fix false (?) buffer overflow
From scan-build: size argument is greater than the free space in the
destination buffer, for the line with strncat().

This might be a false positive in scan-build, the size looks right
at first glance. To be on the safe side and get rid of the warning,
allocate one one more byte...