D-Bus testing: fix and enhance TestHTTP
Picking a port dynamically was broken: if 9999 was in use, the test script
connected to the wrong process and then incorrectly continued, instead of
detecting that syncevo-http-server failed. Use the auto-alloc feature
of syncevo-http-server (port == 0) instead and get the actual port
from the output.

Now that we redirect that output into a file, it makes sense to also include
that log in test failure reports.