autotools: build issue with PIM Manager + boost-locale
On current Debian Testing, boost libraries are installed in /usr/lib/<arch>
directories where they are not found by the Boost m4 macros. This breaks the
boost-locale detection because it only tries linking against libs that it
finds in the file system first.

To fix this, at least try linking with -lboost-locale and use that if it
works.