testing: wait for background daemon in wrappercheck
When starting a daemon via wrappercheck, optionally wait for the
daemon to write a certain message to its log file. The message is
searched for with "grep -e", so regular expressions matching one line
are acceptable. Only works in combination with writing that log file.

This will be used to wait for syncevo-http-server to confirm that it
is running and ready to accept connections.