testing: wait for syncevo-http-server, enable more logging
Use the new wrappercheck.sh option --wait-for-daemon-output to ensure
that the syncevo-http-server is really running. This solves both a
timing issue (tests might start too soon) and make it more obvious when
the HTTP server is not working properly.

Diagnosing that was hard. Now we detect it and also write more debug
output from the HTTP server side into the syncevohttp.log.