testing: replace CT_ASSERT_TRUE
The macro was supposed to help Clang scan-build detect that error
paths are not taken. However, it caused the expression to be
evaluated twice, which is not okay for some more recent checks.

Instead we now enable either the real assert or our simplified check,
depending on a custom CT_ASSERT_SIMPLE that has to be set when running with
scan-build.

The same applies to "no throw" asserts. To keep the wrapper macros
simple, we pass the same statement twice.